+++ dao/userDao.js
@@ -1,19 +1,6 @@
-import path from 'path'
-import { Low, JSONFile } from 'lowdb'
+import db from '../db.js';
 
-const __dirname = path.resolve();
-
-// Use JSON file for storage
-const file = path.join(__dirname, '/db.json')
-const adapter = new JSONFile(file)
-const db = new Low(adapter)
-
-// init
-await db.read()
-db.data = db.data || { users: [] } // Node < v15.x
-await db.write()
-
-const registUser = (newUser) => {
+const insertUser = (newUser) => {
     db.data.users.push(newUser);
     db.write();
 }
@@ -32,8 +19,8 @@ const getUserByEmail = (email) => {
     return null;
 }
 
-export {
+export default {
     getAllUsers,
-    registUser,
+    insertUser,
     getUserByEmail,
 };
